Seperated Neo4j Clients for Graph and Analysis Engines
