+++ sonar/analyze/neo4j_analysis_client.py
@@ -1,0 +1,5 @@
+from neo4j_client import Neo4jClient
+
+class Neo4jAnalysisClient(Neo4jClient):
+
+    pass
