+++ sonar/analyze/views.py
@@ -1,3 +1,9 @@
-from django.shortcuts import render
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+class AnalysisView(APIView):
+
+    def post(self, request):
+        return Response({"message": "Hello, world!"})
 
 # Create your views here.
